Ubiquity now only regenerates sandboxes when code sources have changed, which dramatically improves Ubiquity's responsiveness when the user enters Ubiquity.
Note, however, that this means command sandboxes won't be regenerated every time a command has been run, which can give the illusion that variables in global scope persist between command invocations.  It's still the case that modifying variables in global scope results in undefined behavior--previously it happened to be the case that such globals were reset every time the command was executed, but now it really is unpredictable.  So commands should still use the 'globals' object or the 'windowGlobals' object to store globals that persist within the same firefox session or window, respectively.