Only one instance of each Ubiquity command feed is now loaded for the entire Mozilla application: previously, a separate sandboxed instance was loaded for each browser window, which made creating new windows in Ubiquity extremely slow.
This is actually a fairly big change right now, because it means that the 'window' global no longer exists in the context of Ubiquity command feeds.  Some of the non-window-specific globals like XMLHttpRequest are now retrieved from the hidden window, and jQuery and TrimPath are now loaded in the context of the hidden window and given to sandboxes.  The 'window' global can still be made to exist if we really want it to, but I don't think there's any way that the exact behavior of some currently-existing Ubiquity commands can be preserved.

At the moment, a lot of things are broken: the command list page almost certainly is, and page load functions won't currently load either, nor will functions that are called on ubiquity startup (this is partly because final.js isn't being loaded right now).  I'll be fixing this stuff shortly.

For now, though, the basic search commands like google, youtube, wiki, imdb, etc. still work.  And creating a new browser window is fast again.

--HG--
rename : ubiquity/chrome/content/builtinfactories.js => ubiquity/modules/builtinfactories.js