A usability update.
The contents of the selection you are using Ubiquity to act upon used to not be visually distinguished from the rest of the command.
The selection is now styled according to the .selected selector of the CSS skin. This enables a visual understanding of exactly
what the command is doing, and where the pronoun subsitution is taking place.

I should note, however that this is implemented with something of a hack. Right now, there is no good way to mark where
pronoun subsitution occured. Ideally, this would be done with some metadata attached to the suggestion string that
could be later used to markup that suggestion with HTML tags. However, in JS, strings can't have metadata attached to them.
In the future, this means that the suggestion text should be passed around as an object. For now, however, the metadata
is being embedded using some un-used unicode characters directy into the string and stripped out later.

TODO: Fix this hack. After having gone through making this change it appears that the changes need only be made to
* _argSuggest() in parser.js
* things that call _markSelectionHACK() in parser.js
* NounUtils.makeSugg in nounutils.js