Added dictionary test, as well as a few more bells and whistles to the test framework.
The jQuery tabs extension was somehow messing things up so that Firebug's console re-injected itself into the page after we installed our custom console handler. Addtionally, the new tab structure had broken up the page such that the orange warnings weren't showing when they were supposed to, and broke up information so that Jetpack was harder to debug than it used to be. As such, I've reverted back to the old single-page design, which fixes the console issue; we may want to consider just separating the user-facing page from the developer dashboard/jetpack runtime entirely.