This is a commit of changeset c37dbc723b9f from the ubiquity 0.1.9 branch:
  https://ubiquity.mozilla.com/hg/ubiquity-firefox-0.1.9/rev/c37dbc723b9f

Here's the changelog from that commit:

Added a button to the reset page allowing users to export their current ubiquity configuration to a new tab, in JSON format. This can eventually be used to figure out what in the ubiquity configuration was causing ubiquity to crash; it can also allow users to manually recover some of their data after a reset.